+++ src/elexmodel/client.py
@@ -37,7 +37,6 @@ class ModelClient:
 
     def __init__(self):
         super().__init__()
-        self.model = None
         self.all_conformalization_data_unit_dict = defaultdict(dict)
         self.all_conformalization_data_agg_dict = defaultdict(dict)
 
@@ -223,13 +222,13 @@ class ModelClient:
         )
 
         if pi_method == "nonparametric":
-            self.model = NonparametricElectionModel(model_settings=model_settings)
+            model = NonparametricElectionModel(model_settings=model_settings)
         elif pi_method == "gaussian":
-            self.model = GaussianElectionModel(model_settings=model_settings)
+            model = GaussianElectionModel(model_settings=model_settings)
 
         minimum_reporting_units_max = 0
         for alpha in prediction_intervals:
-            minimum_reporting_units = self.model.get_minimum_reporting_units(alpha)
+            minimum_reporting_units = model.get_minimum_reporting_units(alpha)
             if minimum_reporting_units > minimum_reporting_units_max:
                 minimum_reporting_units_max = minimum_reporting_units
 
@@ -261,22 +260,22 @@ class ModelClient:
         )
 
         for estimand in estimands:
-            unit_predictions = self.model.get_unit_predictions(reporting_units, nonreporting_units, estimand)
+            unit_predictions = model.get_unit_predictions(reporting_units, nonreporting_units, estimand)
             results_handler.add_unit_predictions(estimand, unit_predictions)
             # gets prediciton intervals for each alpha
             alpha_to_unit_prediction_intervals = {}
             for alpha in prediction_intervals:
-                alpha_to_unit_prediction_intervals[alpha] = self.model.get_unit_prediction_intervals(
+                alpha_to_unit_prediction_intervals[alpha] = model.get_unit_prediction_intervals(
                     results_handler.reporting_units, results_handler.nonreporting_units, alpha, estimand
                 )
-                if isinstance(self.model, ConformalElectionModel): 
-                    self.all_conformalization_data_unit_dict[alpha][estimand] = self.model.get_all_conformalization_data_unit()
+                if isinstance(model, ConformalElectionModel): 
+                    self.all_conformalization_data_unit_dict[alpha][estimand] = model.get_all_conformalization_data_unit()
             
             results_handler.add_unit_intervals(estimand, alpha_to_unit_prediction_intervals)
 
             for aggregate in results_handler.aggregates:
                 aggregate_list = self.get_aggregate_list(office, aggregate)
-                estimates_df = self.model.get_aggregate_predictions(
+                estimates_df = model.get_aggregate_predictions(
                     results_handler.reporting_units,
                     results_handler.nonreporting_units,
                     results_handler.unexpected_units,
@@ -285,7 +284,7 @@ class ModelClient:
                 )
                 alpha_to_agg_prediction_intervals = {}
                 for alpha in prediction_intervals:
-                    alpha_to_agg_prediction_intervals[alpha] = self.model.get_aggregate_prediction_intervals(
+                    alpha_to_agg_prediction_intervals[alpha] = model.get_aggregate_prediction_intervals(
                         results_handler.reporting_units,
                         results_handler.nonreporting_units,
                         results_handler.unexpected_units,
@@ -294,8 +293,8 @@ class ModelClient:
                         alpha_to_unit_prediction_intervals[alpha],
                         estimand,
                     )
-                    if isinstance(self.model, ConformalElectionModel):
-                        self.all_conformalization_data_agg_dict[alpha][estimand] = self.model.get_all_conformalization_data_agg()
+                    if isinstance(model, ConformalElectionModel):
+                        self.all_conformalization_data_agg_dict[alpha][estimand] = model.get_all_conformalization_data_agg()
 
                 # get all of the prediction intervals here
                 results_handler.add_agg_predictions(
