+++ tests/models/test_gaussian_election_model.py
@@ -2,14 +2,16 @@ from elexmodel.models import GaussianElectionModel
 
 
 def test_compute_conf_frac():
-    model = GaussianElectionModel.GaussianElectionModel()
+    model_settings = {}
+    model = GaussianElectionModel.GaussianElectionModel(model_settings)
     conf_frac = model._compute_conf_frac()
 
     assert conf_frac == 0.7
 
 
 def test_get_minimum_reporting_units():
-    model = GaussianElectionModel.GaussianElectionModel()
+    model_settings = {}
+    model = GaussianElectionModel.GaussianElectionModel(model_settings)
     n_min = model.get_minimum_reporting_units(0.7)
 
     assert n_min == 7
